100 - move find unique graphs to sql_data_holder.py file
