+++ tel2puml/otel_to_pv/data_holders/base.py
@@ -96,3 +96,34 @@ class DataHolder(ABC):
         :rtype: `dict`[`str`, `set`[`str`]]
         """
         pass
+
+
+def get_time_window(
+    time_buffer: int, data_holder: DataHolder
+) -> tuple[int, int]:
+    """Get the time window for the unique graphs.
+
+    :param time_buffer: The time buffer to add to the time window in minutes
+    :type time_buffer: `int`
+    :param data_holder: The data holder object containing the ingested data
+    :type data_holder: :class:`DataHolder`
+    :return: The time window for the unique graphs
+    :rtype: `tuple`[`int`, `int`]
+    :raises ValueError: If the time buffer is too large for the ingested data
+    """
+    time_buffer_in_nanoseconds = time_buffer * 60 * 1000000000
+    buffer_min_timestamp = (
+        data_holder.min_timestamp + time_buffer_in_nanoseconds
+    )
+    buffer_max_timestamp = (
+        data_holder.max_timestamp - time_buffer_in_nanoseconds
+    )
+    if buffer_min_timestamp >= buffer_max_timestamp:
+        raise ValueError(
+            "The time buffer is too large for the ingested data. "
+            "Please reduce the time buffer."
+        )
+    return (
+        data_holder.min_timestamp + time_buffer_in_nanoseconds,
+        data_holder.max_timestamp - time_buffer_in_nanoseconds,
+    )
